Remove Papa Dependency
Remove the papa dependency, change any required functions to be explicit
imports. The `lookup` function from papa has been replaced by the one from the
prelude, which is specialised to list, but the use site is referring to a list
anyway.

The list append and concat functions have been replaced by their
semigroup/monoid counterparts.

Update default.nix to clean up the package as the versions have made it into the
all-cabal-hashes.